updated the menus (got rid of metitation toggle since we are not using it). some guardian behavior set up for coral planet branching. a bit of work on pearls.
