+++ Assets/Scripts/AI Behaviors/GuardianBehaviorScriptable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 873e6aafdf8bb7c46968b8a0a3470951
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
