+++ Assets/Scripts/AI Behaviors/GuardianBehaviors.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6300f2746f5db41418c761ccd28f3374
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
